Liberate builtin types from instances
